Add network process test
